requirements.txt
google-genai
python-dotenv
openai
ollama